requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
prisma==0.11.0
pandas==2.1.4
numpy==1.26.0
openpyxl==3.1.2
xlrd==2.0.1
aiofiles==23.2.1
Pillow==10.1.0
psycopg2-binary==2.9.9
psutil==5.9.6
